requirements.txt
openpyxl
pdfplumber
pytesseract
pdf2image
pillow
imapclient
pyzmail
tkcalendar
APScheduler
flask
flask-cors
